SQLの型修正 と MapartManager$getWorldとgenerateWorldの動作変更

+++ doma/src/main/java/net/akazukin/mapart/doma/dto/MapartLandDto.java
@@ -6,8 +6,8 @@ import lombok.Data;
 
 @Data
 public class MapartLandDto {
-    private int landId;
-    private int locationId;
+    private long landId;
+    private long locationId;
 
     private long size;
